Bug fix: deassert the reset signal @ posedge + delay.
Deasserting the signal at posedge enables a weird bug, triggering an
exception in the same moment.

Deasserting before, due to the dual clock testbench, the memory will
output the data in the positive clock region, and high-z in the
negative/zero region.

Solution (TODO): use a single clock, or sync the reset inside the core.